requirements.txt
# =============================================================================
# SNOWFLAKE EYECARE ANALYTICS PLATFORM - PYTHON DEPENDENCIES
# =============================================================================

# -----------------------------------------------------------------------------
# CORE STREAMLIT & WEB FRAMEWORK
# -----------------------------------------------------------------------------
streamlit>=1.28.0

# -----------------------------------------------------------------------------
# CORE SNOWFLAKE & DATABASE
# -----------------------------------------------------------------------------
snowflake-connector-python[pandas]==3.15.0
sqlalchemy==2.0.23
pymssql==2.2.8  # SQL Server (jTDS compatible)

# -----------------------------------------------------------------------------
# SECURITY & AUTHENTICATION
# -----------------------------------------------------------------------------
cryptography>=41.0.0
secrets
hashlib

# -----------------------------------------------------------------------------
# DATA PROCESSING & ANALYTICS
# -----------------------------------------------------------------------------
pandas==2.1.4
numpy==1.24.4
pyarrow==14.0.2
polars==0.20.2  # High-performance data processing
duckdb==0.9.2   # In-memory analytics

# -----------------------------------------------------------------------------
# WEB APPLICATIONS & DASHBOARDS
# -----------------------------------------------------------------------------
streamlit==1.28.2
plotly==5.17.0
altair==5.2.0
bokeh==3.3.2
dash==2.16.1

# -----------------------------------------------------------------------------
# API & WEB FRAMEWORK
# -----------------------------------------------------------------------------
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.2
requests==2.31.0

# -----------------------------------------------------------------------------
# WORKFLOW & ORCHESTRATION
# -----------------------------------------------------------------------------
apache-airflow==2.8.0
prefect==2.14.11
schedule==1.2.0

# -----------------------------------------------------------------------------
# DATA QUALITY & VALIDATION
# -----------------------------------------------------------------------------
great-expectations==0.18.8
pydantic-extra-types==2.2.0
cerberus==1.3.5

# -----------------------------------------------------------------------------
# MACHINE LEARNING & AI
# -----------------------------------------------------------------------------
scikit-learn==1.3.2
lightgbm==4.1.0
xgboost==2.0.2
prophet==1.1.5  # Time series forecasting
openai==1.6.1   # For AI integrations

# -----------------------------------------------------------------------------
# UTILITIES & CONFIGURATION
# -----------------------------------------------------------------------------
python-dotenv==1.0.0
pyyaml==6.0.1
click==8.1.7
rich==13.7.0    # Beautiful terminal output
typer==0.9.0    # CLI framework

# -----------------------------------------------------------------------------
# LOGGING & MONITORING
# -----------------------------------------------------------------------------
loguru==0.7.2
prometheus-client==0.19.0
sentry-sdk==1.40.0

# -----------------------------------------------------------------------------
# TESTING FRAMEWORK
# -----------------------------------------------------------------------------
pytest==7.4.3
pytest-cov==4.1.0
pytest-mock==3.12.0
pytest-asyncio==0.21.1

# -----------------------------------------------------------------------------
# DEVELOPMENT TOOLS
# -----------------------------------------------------------------------------
black==23.11.0      # Code formatting
isort==5.12.0       # Import sorting
flake8==6.1.0       # Linting
mypy==1.7.1         # Type checking
pre-commit==3.6.0   # Git hooks

# -----------------------------------------------------------------------------
# SECURITY & ENCRYPTION
# -----------------------------------------------------------------------------
cryptography==41.0.8
bcrypt==4.1.2
PyJWT==2.8.0

# -----------------------------------------------------------------------------
# CLOUD & INFRASTRUCTURE
# -----------------------------------------------------------------------------
boto3==1.34.0       # AWS SDK
azure-storage-blob==12.19.0  # Azure
google-cloud-storage==2.10.0  # GCP

# -----------------------------------------------------------------------------
# DATA VISUALIZATION & REPORTING  
# -----------------------------------------------------------------------------
matplotlib==3.8.2
seaborn==0.13.0
reportlab==4.0.7    # PDF generation
openpyxl==3.1.2     # Excel files
xlsxwriter==3.1.9

# -----------------------------------------------------------------------------
# NOTIFICATION & COMMUNICATION
# -----------------------------------------------------------------------------
slack-sdk==3.26.1
sendgrid==6.11.0
twilio==8.11.1

# -----------------------------------------------------------------------------
# DEVELOPMENT ENVIRONMENT
# -----------------------------------------------------------------------------
jupyter==1.0.0
ipykernel==6.27.1
notebook==7.0.6
